<Feat> :  MySQL ---------------------------------------- 1. db 함수 폴더 구분 2. query 실행 함수 폴더 구분 ---------------------------------------- 1. 게시판에서 글생성 시 db 생성 2. 댓글 추가 시 db 생성
